requirements.txt
asgiref==3.5.2
backports.entry-points-selectable==1.1.0
cffi==1.15.1
cfgv==3.3.1
cryptography==37.0.4
distlib==0.3.3
Django==4.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==14.1.0
filelock==3.3.1
identify==2.3.0
nodeenv==1.6.0
platformdirs==2.4.0
pre-commit==2.15.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.8.1
psycopg2>=2.8
